# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Classroom management backend (Go net/http ServeMux). Auth via JWT Bearer (HS256, 24h); mini-program identity resolved server-side via code2Session.",
        "title": "OCM 智慧教室管理平台 API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/repairs": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "List repair tickets",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "filter by classroom id",
                        "name": "classroom_id",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "open",
                            "processing",
                            "completed",
                            "confirmed"
                        ],
                        "type": "string",
                        "description": "filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "search",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1-based page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 100,
                        "description": "page size",
                        "name": "page_size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "paged repair views",
                        "schema": {
                            "$ref": "#/definitions/httpx.Paged"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "Create a repair ticket",
                "parameters": [
                    {
                        "description": "repair input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created repair view",
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairView"
                        }
                    },
                    "400": {
                        "description": "invalid body / classroom not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "open ticket already exists for the classroom",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repairs/emergency": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "Create an emergency repair ticket (skips the open-ticket guard)",
                "parameters": [
                    {
                        "description": "repair input",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "created repair view",
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairView"
                        }
                    },
                    "400": {
                        "description": "invalid body / classroom not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repairs/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "Get a repair ticket",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "repair id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "repair detail",
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairView"
                        }
                    },
                    "400": {
                        "description": "invalid repair id",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "repair ticket not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "Process a repair ticket (start/finish)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "repair id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "status transition",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairUpdateInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "updated repair view",
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairView"
                        }
                    },
                    "400": {
                        "description": "invalid body / invalid status transition",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "repair ticket not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/repairs/{id}/confirm": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "repairs"
                ],
                "summary": "Confirm a completed repair (creator only)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "repair id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "confirmed repair view",
                        "schema": {
                            "$ref": "#/definitions/classroom.RepairView"
                        }
                    },
                    "400": {
                        "description": "repair not completed yet",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "only the creator may confirm",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "repair ticket not found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "classroom.RepairInput": {
            "type": "object",
            "required": [
                "classroomId",
                "description",
                "images"
            ],
            "properties": {
                "classroomId": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "classroom.RepairUpdateInput": {
            "type": "object",
            "required": [
                "remark",
                "status"
            ],
            "properties": {
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "classroom.RepairView": {
            "type": "object",
            "required": [
                "assigneeName",
                "building",
                "classroomId",
                "classroomName",
                "createdAt",
                "creatorId",
                "creatorName",
                "description",
                "id",
                "images",
                "remark",
                "status",
                "updatedAt"
            ],
            "properties": {
                "assigneeId": {
                    "type": "integer"
                },
                "assigneeName": {
                    "type": "string"
                },
                "building": {
                    "type": "string"
                },
                "classroomId": {
                    "type": "integer"
                },
                "classroomName": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "creatorId": {
                    "type": "integer"
                },
                "creatorName": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "images": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "remark": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "httpx.ErrorResponse": {
            "type": "object",
            "required": [
                "error"
            ],
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
        "httpx.Paged": {
            "type": "object",
            "required": [
                "items",
                "page",
                "pageSize",
                "total"
            ],
            "properties": {
                "items": {},
                "page": {
                    "type": "integer"
                },
                "pageSize": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "\"JWT Bearer token; prefix with 'Bearer '.\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}